+++ tests/test_multiprocessing_serverprocess.py
@@ -10,6 +10,7 @@ from mpservice.multiprocessing import (
     Queue,
     ServerProcess,
     SpawnProcess,
+    MemoryBlock,
 )
 from mpservice.threading import Thread
 
@@ -230,3 +231,19 @@ def test_shared_memory():
         p.join()
 
         assert len(manager.list_memory_blocks()) == 0
+
+
+class MemoryWorker:
+    def memory_block(self, size):
+        return MemoryBlock(size)
+    
+
+def test_shared_memory_from_serverprocess():
+    ServerProcess.register(MemoryWorker, method_to_typeid={'memory_block': 'memory_block_in_server'})
+    with ServerProcess() as memory_server:
+        with ServerProcess() as server:
+            worker = server.MemoryWorker()
+            mem = worker.memory_block(20)
+            print(type(mem))
+            del mem
+